Split tag cell editor from designer panel
